refactor(task_runner): split into submodules (re-#625)
task_runner.rs was 3,635 lines, exceeding the 800-line VibeGuard U-16
limit. Split into six focused modules following the task_executor/ pattern:

- types.rs      — all type definitions and their impls (~620 lines)
- helpers.rs    — utility functions (transient-error detection, project
                   resolution, prompt planning heuristic) (~330 lines)
- store.rs      — TaskStore struct and all cache/DB methods (~730 lines)
- store_tests.rs — TaskStore tests via #[path] to preserve private access
- recovery.rs   — validate_recovered_tasks (startup PR-state check)
- lifecycle.rs  — spawn_task, register_pending_task, spawn_preregistered_task,
                   update_status, mutate_and_persist (~610 lines)
- lifecycle_tests.rs — spawn/phase/interceptor tests via #[path]
- dependencies.rs — detect_cycle, spawn_task_awaiting_deps, check_awaiting_deps
- mod.rs        — module declarations + pub/pub(crate) re-exports

Public API surface is byte-identical via pub use re-exports.
Zero behaviour change; pure mechanical split.

Refs #625
